Feature: sync recent updates.

+++ plugins/xsettings/ukui-xft-settings.h
@@ -29,7 +29,7 @@ private:
         gboolean    hinting;
         int         dpi;
         int         scaled_dpi;
-        int         window_scale;
+        double      window_scale;
         char       *cursor_theme;
         int         cursor_size;
         const char *rgba;
